master: (1) light

+++ PlayRoom/GameDataSender.h
@@ -18,5 +18,5 @@ extern NSString * const GameDataSenderMatrixUUID;
 
 
 @interface GameDataSender : NSObject
-- (void)sendMatrix:(const CGFloat[16])mat;
+- (void)sendMatrix:(const float[16])mat;
 @end
